v2.4.1 + Working saved JSON file

+++ src/time_work.rs
@@ -6,14 +6,12 @@ pub fn time_parser(s: String) -> i64 {
         now
     } else if s.contains('h') {
         now + chrono::Duration::hours(
-            string_work::extract_number(s.as_str()).expect("Couldn't extract number from query")
-                as i64,
+            string_work::extract_number(s.as_str()).expect("Couldn't extract number from query"),
         )
         .num_seconds()
     } else if s.contains('m') {
         now + chrono::Duration::minutes(
-            string_work::extract_number(s.as_str()).expect("Couldn't extract number from query")
-                as i64,
+            string_work::extract_number(s.as_str()).expect("Couldn't extract number from query"),
         )
         .num_seconds()
     } else if s.contains(';') {
@@ -22,14 +20,14 @@ pub fn time_parser(s: String) -> i64 {
         let m = *a.last().expect("Failed to extract minutes from query");
 
         now + chrono::Duration::hours(
-            string_work::extract_number(h).expect("Couldn't extract number from query") as i64,
+            string_work::extract_number(h).expect("Couldn't extract number from query"),
         )
         .num_seconds()
             + chrono::Duration::minutes(
-                string_work::extract_number(m).expect("Couldn't extract number from query") as i64,
+                string_work::extract_number(m).expect("Couldn't extract number from query"),
             )
             .num_seconds()
     } else {
-        string_work::extract_number(s.as_str()).expect("Couldn't extract number from query") as i64
+        string_work::extract_number(s.as_str()).expect("Couldn't extract number from query")
     }
 }
